+++ packages/chat/src/components/Message/__fixtures__/code-snippet.ts
@@ -1,0 +1,17 @@
+export const CODE_SNIPPET_FIXTURE = `
+\`\`\`javascript
+<script type="text/javascript">
+  (function(d, t) {
+      var v = d.createElement(t), s = d.getElementsByTagName(t)[0];
+      v.onload = function() {
+        window.voiceflow.chat.load({
+          verify: { projectID: 'howdy folks, very cool verification hash here' },
+          url: 'https://general-runtime.voiceflow.com',
+          versionID: 'production'
+        });
+      }
+      v.src = "https://cdn.voiceflow.com/widget/bundle.mjs"; v.type = "text/javascript"; s.parentNode.insertBefore(v, s);
+  })(document, 'script');
+</script>
+\`\`\`
+`;
